requirements.txt
fastapi
uvicorn[standard]
redis
celery
elasticsearch==8.11.1
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
pydantic-settings
requests
python-dotenv
sqlalchemy
psycopg2-binary
PyPDF2
python-docx
google-generativeai>=0.3.0
sentence-transformers
numpy==1.26.4
